ユニコの脳みそ/00/03_プロジェクト資料/Legacy_Tools/Research_Business_Tool/requirements.txt
openai
duckduckgo-search
jinja2
python-dotenv
beautifulsoup4
requests
streamlit
youtube-transcript-api
